Fix invalid API design
From the challenge summary:

> Implement a prototype job worker service that provides an API to run **arbitrary Linux processes**.

Emphasis mine.

I had missed that part when I first did my design, so this commit is fixing up
the API design so that it's properly addressing the challenge.

The main difference is the "arbitrary" part, it just means we can't rigidly
define the arguments for a job as part of the API.