Fixing up magefiles
Switched to using `mg.SerialDeps` instead of `mg.Deps` in a few places because
some of the dependencies can't be run in parallel.

Cleaned up the output that happens when mage is called. Used a buffer for STDOUT
and another for STDERR when running a command, so that I can control *when* the
output from a command is seen. Doing this so that the output is hidden if the
command is a success, so that the output is cleaner. Don't need to see a bunch
of verbose output when nothing went wrong.