Cleaning up code a bit more
Removed testDial from the client tests because it wasn't really needed.

Finished the test server for use in the client tests.

The client checks the id given now; this is in *addition* to the checking done
on the server side ( and in the library ). Checking it in the client means we're
not wasting bandwidth on invalid (some) requests; still need to check in the
server & library because there might be non-Go clients or servers(?) that don't
check the ID properly.