Type alias some GRPC stuff, implement authorization
For the type aliases: was running into a circular-import issue, this was the
quick fix. If I had more time I'd refactor things so that we could avoid the
circular import, but I've got a time budget and can't spend too much time on
this.

As for the authorization stuff: it works great! The system handles incoming
requests properly, and users can only see their own jobs -- unless they've got
the `Super` privilege, then they can do what they want.