Re-write server binary as cobra command
So that we can get flags, as well as start prepping for the chicanery we have to
get up to in order to do the cgroup/namespace stuff.

Also setup two methods in `library/api/startup.go` that are called in the server
binary main(). Both are 'special' commands, used while launching a job. They're
also hidden from the 'help' output -- right now, this is what gets printed when
you run the server binary with no arguments:

```
The GRPC server for workernator

Usage:
  workernator [command]

Available Commands:
  completion                  Generate the autocompletion script for the specified shell
  help                        Help about any command
  run                         Starts the GRPC server

Flags:
  -h, --help   help for workernator

Use "workernator [command] --help" for more information about a command.
```

As you can see, neither of the special commands show up. You'd be able to tell,
because in the `Available Commands` section you'd see these two commands:

 - `__WORKERNATOR__NAMESPACED__`
 - `__WORKERNATOR__FINAL__`

Neat!