Implement a few concurrency fixes/tweaks
First, moved some of the code in `StartJob` into their own functions to make it
a bit easier to read. One of the functions just cares about the STDOUT and
STDERR files; once the command is done it closes those files. The other waits
for the job to finish, then handles updating the job data & canceling the
context.

Another change is switching the type returned form `GetJobOutput` from
`io.Reader` to `io.ReadCloser`. This is so that the pipe reader returned by this
method can be closed; this is how the goroutine reading from the file knows it
needs to close exit. In tandem with this, updated `(*Service).Output)` ( the
GRPC handler ) so that it launches a goroutine that waits for the context to
signal it's done; when it does it calls `Close()` on the reader returned from
the library.

Moved the `statusKilled` constant from `library/api` to `library`, and also
added a `statusOK` constant. Moved here because I moved some of the "set some
stuff in the job struct" into a method on the job itself; the constants just
followed the cleanup code. This new function `SetFinished(cmdErr error, exitCode
int)` uses a lock to ensure that there's no tricky race conditions when trying
to find out of a job has finished or not.